requirements.txt
requests
python-dotenv
beautifulsoup4       # HTML解析用
scikit-learn>=1.3.0  # TF-IDF分析用
janome>=0.5.0        # 日本語形態素解析用
